Add confirm option to install rails scaffold

+++ jefe-cli.sh
@@ -484,34 +484,37 @@ configure_ruby_project() {
 
     create_folder_structure
 
-    out "Select type of project ruby" 5
-    out "0) Ripper Rails API Project" 5
-    echo "Type the option (number) that you want(digit), followed by [ENTER]:"
-    read option
-    flag=true
-    while [ $flag = true ]; do
-        case $option in
-            0)
-                cd ./${project_root}
-                git clone https://github.com/ricardo-pcan/ripper-rails-scaffold.git .
-                rm -rf .git
-                cp .circle.yml ../circle.yml
-                rm .circle.yml
-                rm -f ../.gitignore
-                cp .gitignore ../.gitignore
-                rm .gitignore
-                cp .env.example .env
-                cd ..
-                flag=false
-                ;;
-            *)
-                out "Wrong choice:$option" 1
-                project=""
-                flag=true
-                ;;
-        esac
-    done
-    cp .jefe/postinstall.sh ./$project_root/postinstall.sh
+    out "Create new proyect?"
+    if [[ $REPLY =~ ^[Yy]$ ]]; then
+        out "Select type of project ruby" 5
+        out "0) Ripper Rails API Project" 5
+        echo "Type the option (number) that you want(digit), followed by [ENTER]:"
+        read option
+        flag=true
+        while [ $flag = true ]; do
+            case $option in
+                0)
+                    cd ./${project_root}
+                    git clone https://github.com/ricardo-pcan/ripper-rails-scaffold.git .
+                    rm -rf .git
+                    cp .circle.yml ../circle.yml
+                    rm .circle.yml
+                    rm -f ../.gitignore
+                    cp .gitignore ../.gitignore
+                    rm .gitignore
+                    cp .env.example .env
+                    cd ..
+                    flag=false
+                    ;;
+                *)
+                    out "Wrong choice:$option" 1
+                    project=""
+                    flag=true
+                    ;;
+            esac
+        done
+        cp .jefe/postinstall.sh ./$project_root/postinstall.sh
+    fi
 }
 
 # Docker compose var env configuration.
